Port control via nat-api
