+++ package.json
@@ -27,8 +27,8 @@
     "dts-gen": "^0.6.0",
     "elliptic": "^6.5.4",
     "express": "^4.17.1",
-    "freedom-port-control": "^0.9.11",
     "hjson": "^3.2.2",
+    "nat-api": "^0.3.1",
     "nat-upnp-wrapper": "^1.5.0",
     "node": "^16.13.1",
     "npm": "^8.3.1"
